requirements.txt
requests
doxypypy
jellyfish
lxml
matplotlib
mkdocs
scipy
networkx
numpy
pandas
pylint
selenium
tabulate
gitpython
html5lib
beautifulsoup4
